✨ feat(input): внедрена событийная архитектура пользовательского ввода
🚀 Добавлены ключевые классы для новой модели обработки действий:
• Action — базовый абстрактный тип пользовательского действия
• ActionContext — универсальный контейнер параметров для передачи в обработчики
• ActionRegistry — централизованный диспетчер, поддерживающий регистрацию и вызов множественных обработчиков по имени действия

🔧 Эта система вводит push-модель взаимодействия, заменяя прежнюю систему на основе опроса состояния клавиш.
Обеспечивает расширяемую и независимую обработку ввода, готовую к интеграции с новыми источниками (геймпады, UI, AI и т.д.).

📦 Пакет: pyc.lopatuxin.core.input

+++ src/main/java/pyc/lopatuxin/core/components/movement/VelocityComponent.java
@@ -16,6 +16,6 @@ import pyc.lopatuxin.core.ecs.Component;
 @AllArgsConstructor
 public class VelocityComponent implements Component {
 
-    private float velocityX;
-    private float velocityY;
+    private int velocityX;
+    private int velocityY;
 }
