+++ src/main/java/pyc/lopatuxin/core/entities/PlayerEntity.java
@@ -1,19 +1,18 @@
 package pyc.lopatuxin.core.entities;
 
-import pyc.lopatuxin.core.components.movement.PlayerControlComponent;
 import pyc.lopatuxin.core.components.movement.PositionComponent;
 import pyc.lopatuxin.core.components.movement.VelocityComponent;
 import pyc.lopatuxin.core.ecs.Entity;
 
 public class PlayerEntity extends Entity {
 
-    public PlayerEntity(float startX, float startY) {
+    public PlayerEntity(int startX, int startY) {
         addComponent(new PositionComponent(startX, startY));
         addComponent(new VelocityComponent(0, 0));
         addComponent(new PlayerControlComponent());
     }
 
-    public void setVelocity(float vx, float vy) {
+    public void setVelocity(int vx, int vy) {
         VelocityComponent velocity = getComponent(VelocityComponent.class);
         if (velocity != null) {
             velocity.setVelocityX(vx);
